+++ CHANGELOG.md
@@ -31,8 +31,11 @@ of requirements for an API to count as public.
 ### Features
 
 - *Breaking*: `get_jwt` is renamed to `request_jwt`, #868
+- *Breaking*: `ResponseSpecProvider.provide_response_specs` is now
+  an instance method, #877
 - Added official PyPy 3.11+ support, #870
-- Now `request.auth` is set on all successful auth workflows, #868
+- Added `dmr.throttling` package, #877
+- Added `request.__drm_auth__` on all successful auth workflows, #868
 - Added `request_auth` helper function, #868
 - Added `strict` parameter to `request_renderer` and `request_parser`,
   added `@overload`s to both of these functions, #869
@@ -45,7 +48,7 @@ of requirements for an API to count as public.
 
 ### Fixes
 
-- Fixed that `OpenAPI` was revalidate on every `.convert` call, #867
+- Fixed that `OpenAPI` was revalidated on every `.convert` call, #867
 
 ### Misc